Fix UI freeze from PosTap render-object and error swallowing
Replace the custom RenderPosTap (which corrupted pointer routing and
caused null-check / mouse_tracker storms) with a plain GestureDetector.
Remove FlutterError filtering that left mouse tracker state inconsistent
and made the cashier extremely slow until disconnect.

+++ lib/core/widgets/hasim_widgets.dart
@@ -396,8 +396,10 @@ class ProductCard extends StatelessWidget {
                                     width: w,
                                     fadeInDuration: Duration.zero,
                                     fadeOutDuration: Duration.zero,
-                                    memCacheWidth:
-                                        (w * 2).round().clamp(64, 512),
+                                    memCacheWidth: (w * 2)
+                                        .round()
+                                        .clamp(64, 512)
+                                        .toInt(),
                                     placeholder: (_, _) => const ColoredBox(
                                       color: HasimColors.surfaceSoft,
                                     ),
